Added errorhandler middleware
